requirements.txt
django
djangorestframework==3.12.1
celery
black
psycopg2-binary
